Add logic to deleteItems to check for user_id

+++ application.py
@@ -379,6 +379,7 @@ def deleteItems(category_id, item_id):
     DBSession = sessionmaker(bind=engine)
     session = DBSession()
     item = session.query(Item).filter_by(id=item_id, category_id=category_id).one()
+    creator = getUserInfo(item.user_id)
     if "username" not in login_session:
         return redirect('/login')
     if request.method == 'POST':
@@ -387,6 +388,9 @@ def deleteItems(category_id, item_id):
         flash('Deleted Requested Item')
         return redirect(url_for('showItems', category_id=category_id))
     else:
+        if creator.id != login_session['user_id']:
+            flash('You are not the owner of the category, Hence you are not allowed to delete the item.')
+            return redirect(url_for('showItems', category_id=category_id))
         return render_template('deleteItem.html', category_id=category_id, item=item)
